requirements.txt
fastapi
uvicorn

passlib
python-jose[cryptography]
python-multipart
mongoengine